Create provider_institutions
